Enabling MinGW compilation.
- Still working on MacOS

Requires:

git submodule update --init
pacboy -S gmp:p eigen3:p cgal:p opencl-headers:p opencl-icd:p freeimage:p wxwidgets3.2-msw:p arpack:p superlu:p blas:p nvidia-cg-toolkit:p mesa:p
cmake -G "MSYS Makefiles" -DWITH_LIB3DS:boolean=OFF -DWITH_CLUTO:boolean=OFF -DWITH_PLUGIN_ARPACK:boolean=OFF ..

+++ Code/Source/Algorithms/DiscreteExponentialMap.cpp
@@ -121,7 +121,7 @@ class Impl
           // Blend in the positions of nearby upwind points (the predecessor's visited neighbors) to make the estimate a bit
           // more robust
           auto const & pred_nbrs = pred->getNeighbors();
-          for (int i = 0; i < pred_nbrs.size(); ++i)
+          for (size_t i = 0; i < pred_nbrs.size(); ++i)
           {
             auto const & pred_nbr = pred_nbrs[i];
             ParamDataMap::const_iterator existing_nbr = param_data.find(pred_nbr.getSample());
